+++ tests/test_webhooks.py
@@ -34,6 +34,7 @@ from crate.operator.webhooks import (
     WebhookInfoChangedPayload,
     WebhookPayload,
     WebhookScalePayload,
+    WebhookSnapshotRestoredPayload,
     WebhookStatus,
     WebhookTemporaryFailurePayload,
     WebhookUpgradePayload,
@@ -59,6 +60,7 @@ def test_payload_serialization_scale():
         health_data=None,
         feedback_data=None,
         backup_schedule_data=None,
+        snapshot_restored_data=None,
     )
     assert json.loads(json.dumps(p)) == {
         "event": "scale",
@@ -78,6 +80,7 @@ def test_payload_serialization_scale():
         "health_data": None,
         "feedback_data": None,
         "backup_schedule_data": None,
+        "snapshot_restored_data": None,
     }
 
 
@@ -97,6 +100,7 @@ def test_payload_serialization_upgrade():
         health_data=None,
         feedback_data=None,
         backup_schedule_data=None,
+        snapshot_restored_data=None,
     )
     assert json.loads(json.dumps(p)) == {
         "event": "upgrade",
@@ -116,6 +120,7 @@ def test_payload_serialization_upgrade():
         "health_data": None,
         "feedback_data": None,
         "backup_schedule_data": None,
+        "snapshot_restored_data": None,
     }
 
 
@@ -199,6 +204,7 @@ class TestWebhookClientSending(AioHTTPTestCase):
                 "health_data": None,
                 "feedback_data": None,
                 "backup_schedule_data": None,
+                "snapshot_restored_data": None,
             },
         }
 
@@ -237,6 +243,7 @@ class TestWebhookClientSending(AioHTTPTestCase):
                 "health_data": None,
                 "feedback_data": None,
                 "backup_schedule_data": None,
+                "snapshot_restored_data": None,
             },
         }
 
@@ -270,6 +277,7 @@ class TestWebhookClientSending(AioHTTPTestCase):
                 "health_data": None,
                 "feedback_data": None,
                 "backup_schedule_data": None,
+                "snapshot_restored_data": None,
             },
         }
 
@@ -301,6 +309,39 @@ class TestWebhookClientSending(AioHTTPTestCase):
                 "health_data": None,
                 "feedback_data": None,
                 "backup_schedule_data": None,
+                "snapshot_restored_data": None,
+            },
+        }
+
+    @unittest_run_loop
+    async def test_send_snapshot_restored_notification(self):
+        response = await self.webhook_client.send_notification(
+            "my-namespace",
+            "my-cluster",
+            WebhookEvent.SNAPSHOT_RESTORED,
+            WebhookSnapshotRestoredPayload(admin_username="admin"),
+            WebhookStatus.SUCCESS,
+            logging.getLogger(__name__),
+        )
+        assert response.status == 200
+        data = await response.json()
+        assert data == {
+            "username": "itsme",
+            "password": "secr3t password",
+            "payload": {
+                "event": "snapshot_restored",
+                "status": "success",
+                "namespace": "my-namespace",
+                "cluster": "my-cluster",
+                "scale_data": None,
+                "upgrade_data": None,
+                "compute_changed_data": None,
+                "info_data": None,
+                "temporary_failure_data": None,
+                "health_data": None,
+                "feedback_data": None,
+                "backup_schedule_data": None,
+                "snapshot_restored_data": {"admin_username": "admin"},
             },
         }
 
@@ -331,6 +372,7 @@ class TestWebhookClientSending(AioHTTPTestCase):
                 "temporary_failure_data": None,
                 "feedback_data": None,
                 "backup_schedule_data": None,
+                "snapshot_restored_data": None,
                 "health_data": {"status": "GREEN"},
             },
         }